requirements.txt
python-dotenv
groq
cohere
requests
transformers
torch